RQLY-515 review-changes: function renames and handle http
